split Text into Text and TextTools
