+++ app/src/main/java/io/github/sspanak/tt9/ime/modes/helpers/AutoSpace.java
@@ -70,7 +70,7 @@ public class AutoSpace {
 
 		return
 			nextKey != 1
-			&& !nextChars.nextIsPunctuation()
+			&& !Text.nextIsPunctuation(nextChars.toString())
 			&& !nextChars.startsWithNumber()
 			&& (
 				previousChar == '.'
